Rewrite app on one Lambda. Rewrite CRUD. Make everything pretty.

+++ routes/userRouter/mainRouter.js
@@ -1,29 +1,35 @@
 const validator = require('../../helper/userValidator');
-const findAllUsers = require('../../controller/userController/findAllUsers');
-const findUserById = require('../../controller/userController/findUserById');
-const createUser = require('../../controller/userController/createUser');
-const updateUser = require('../../controller/userController/updateUser');
-const deleteUser = require('../../controller/userController/deleteUser');
+const findAllUsers = require('./findAllUsersRouter');
+const findUserById = require('./findUserByIdRouter');
+const createUser = require('./createUserRouter');
+const updateUser = require('./updateUserRouter');
+const deleteUser = require('./deleteUserRouter');
+const resp = require('../../helper/responseObject');
 module.exports = (method, url, body) => {
-    const secondUrlPart = url.split('%2')[1];
-    if(url.split('%2')[2]) throw new Error('Please, check entered URL');
+    const secondUrlPart = url.split('%2F')[1];
+    if(url.split('%2F')[2]) return resp(404, {success: false, message: "This URL is not found"});
 
-    if (!secondUrlPart || method === "GET") {
-        const allUsers = findAllUsers();
+
+    if (!secondUrlPart && method === "GET") {
+        return findAllUsers();
     }
-    if (secondUrlPart || method === "GET") {
-        const oneUser = findUserById(secondUrlPart);
+    if (secondUrlPart && method === "GET") {
+        console.log(secondUrlPart);
+        return findUserById(secondUrlPart);
     }
-    if (!secondUrlPart || method === "POST") {
+    if (!secondUrlPart && method === "POST") {
         const validUser = validator(body, method);
-        const isUserCrated = createUser(validUser);
+        return createUser(validUser);
     }
-    if (!secondUrlPart || method === "PUT") {
-        const validUser = validator(body, method)
-        const isUserupdate = updateUser(validUser);
+    if (!secondUrlPart && method === "PUT") {
+        const validUser = validator(body, method);
+        console.log(validUser);
+        return updateUser(validUser);
     }
-    if (!secondUrlPart || method === "DELETE") {
-        const isUserupdate = deleteUser(secondUrlPart);
+    if (!secondUrlPart && method === "DELETE") {
+        return deleteUser(secondUrlPart);
+    }
 
-    }
+
+    return resp(404, {success: false, message: "Please check method"});
 };
